Hub World Camera now officially works!
Re-added the shop and changed its functionality to be inline with how the camera works.

--The Shop--
When the camera is targettting the shop, it now activates the shop panel, once it has stopped targetting the shop the shop panel closes

--TargetCameraScript--
This script allows the user to cycle through (Within array range) from the Shop ----> Boss Room (Including all worlds E, A, W, F) by pressing Q and E to go left and right. PLEASE NOTE: Moving the player in the hub world also moves the camera position and causes the rotation to be iffy.

--Notes--
I can write a seperate script that actives each of the doors (UI Elements) to allow the user to start each world. I personally thinking of clamping movement in the hub world for now. When it does come down to the final boss I will create a seperate scene that is a replica off the hub world but without that camera movement. So as to allow seemless movement of the player.

--Dungeon Map--
I honestly could not figure out how to apply all the box colliders to the map and get it to work correctly. (APOLOGIES!)